+++ internal/provider/model_resource_meraki_appliance_vlans.go
@@ -21,7 +21,6 @@ package provider
 import (
 	"context"
 	"fmt"
-	"maps"
 	"net/url"
 	"slices"
 
@@ -45,34 +44,22 @@ type ResourceApplianceVLANs struct {
 }
 
 type ResourceApplianceVLANsItems struct {
-	Id                     types.String                                        `tfsdk:"id"`
-	ApplianceIp            types.String                                        `tfsdk:"appliance_ip"`
-	Cidr                   types.String                                        `tfsdk:"cidr"`
-	DhcpBootFilename       types.String                                        `tfsdk:"dhcp_boot_filename"`
-	DhcpBootNextServer     types.String                                        `tfsdk:"dhcp_boot_next_server"`
-	DhcpBootOptionsEnabled types.Bool                                          `tfsdk:"dhcp_boot_options_enabled"`
-	DhcpHandling           types.String                                        `tfsdk:"dhcp_handling"`
-	DhcpLeaseTime          types.String                                        `tfsdk:"dhcp_lease_time"`
-	DnsNameservers         types.String                                        `tfsdk:"dns_nameservers"`
-	GroupPolicyId          types.String                                        `tfsdk:"group_policy_id"`
-	VlanId                 types.String                                        `tfsdk:"vlan_id"`
-	Mask                   types.Int64                                         `tfsdk:"mask"`
-	Name                   types.String                                        `tfsdk:"name"`
-	Subnet                 types.String                                        `tfsdk:"subnet"`
-	TemplateVlanType       types.String                                        `tfsdk:"template_vlan_type"`
-	VpnNatSubnet           types.String                                        `tfsdk:"vpn_nat_subnet"`
-	FixedIpAssignments     map[string]ResourceApplianceVLANsFixedIpAssignments `tfsdk:"fixed_ip_assignments"`
-	Ipv6Enabled            types.Bool                                          `tfsdk:"ipv6_enabled"`
-	Ipv6PrefixAssignments  []ResourceApplianceVLANsIpv6PrefixAssignments       `tfsdk:"ipv6_prefix_assignments"`
-	MandatoryDhcpEnabled   types.Bool                                          `tfsdk:"mandatory_dhcp_enabled"`
-	DhcpOptions            []ResourceApplianceVLANsDhcpOptions                 `tfsdk:"dhcp_options"`
-	DhcpRelayServerIps     types.List                                          `tfsdk:"dhcp_relay_server_ips"`
-	ReservedIpRanges       []ResourceApplianceVLANsReservedIpRanges            `tfsdk:"reserved_ip_ranges"`
-}
-
-type ResourceApplianceVLANsFixedIpAssignments struct {
-	Ip   types.String `tfsdk:"ip"`
-	Name types.String `tfsdk:"name"`
+	Id                     types.String                                  `tfsdk:"id"`
+	ApplianceIp            types.String                                  `tfsdk:"appliance_ip"`
+	Cidr                   types.String                                  `tfsdk:"cidr"`
+	DhcpBootOptionsEnabled types.Bool                                    `tfsdk:"dhcp_boot_options_enabled"`
+	DhcpHandling           types.String                                  `tfsdk:"dhcp_handling"`
+	DhcpLeaseTime          types.String                                  `tfsdk:"dhcp_lease_time"`
+	GroupPolicyId          types.String                                  `tfsdk:"group_policy_id"`
+	VlanId                 types.String                                  `tfsdk:"vlan_id"`
+	Mask                   types.Int64                                   `tfsdk:"mask"`
+	Name                   types.String                                  `tfsdk:"name"`
+	Subnet                 types.String                                  `tfsdk:"subnet"`
+	TemplateVlanType       types.String                                  `tfsdk:"template_vlan_type"`
+	Ipv6Enabled            types.Bool                                    `tfsdk:"ipv6_enabled"`
+	Ipv6PrefixAssignments  []ResourceApplianceVLANsIpv6PrefixAssignments `tfsdk:"ipv6_prefix_assignments"`
+	MandatoryDhcpEnabled   types.Bool                                    `tfsdk:"mandatory_dhcp_enabled"`
+	DhcpOptions            []ResourceApplianceVLANsDhcpOptions           `tfsdk:"dhcp_options"`
 }
 
 type ResourceApplianceVLANsIpv6PrefixAssignments struct {
@@ -87,12 +74,6 @@ type ResourceApplianceVLANsDhcpOptions struct {
 	Code  types.String `tfsdk:"code"`
 	Type  types.String `tfsdk:"type"`
 	Value types.String `tfsdk:"value"`
-}
-
-type ResourceApplianceVLANsReservedIpRanges struct {
-	Comment types.String `tfsdk:"comment"`
-	End     types.String `tfsdk:"end"`
-	Start   types.String `tfsdk:"start"`
 }
 
 // End of section. //template:end types
@@ -115,12 +96,6 @@ func (data ResourceApplianceVLANsItems) toBody(ctx context.Context, state Resour
 	if !data.Cidr.IsNull() {
 		body, _ = sjson.Set(body, "cidr", data.Cidr.ValueString())
 	}
-	if !data.DhcpBootFilename.IsNull() {
-		body, _ = sjson.Set(body, "dhcpBootFilename", data.DhcpBootFilename.ValueString())
-	}
-	if !data.DhcpBootNextServer.IsNull() {
-		body, _ = sjson.Set(body, "dhcpBootNextServer", data.DhcpBootNextServer.ValueString())
-	}
 	if !data.DhcpBootOptionsEnabled.IsNull() {
 		body, _ = sjson.Set(body, "dhcpBootOptionsEnabled", data.DhcpBootOptionsEnabled.ValueBool())
 	}
@@ -129,9 +104,6 @@ func (data ResourceApplianceVLANsItems) toBody(ctx context.Context, state Resour
 	}
 	if !data.DhcpLeaseTime.IsNull() {
 		body, _ = sjson.Set(body, "dhcpLeaseTime", data.DhcpLeaseTime.ValueString())
-	}
-	if !data.DnsNameservers.IsNull() {
-		body, _ = sjson.Set(body, "dnsNameservers", data.DnsNameservers.ValueString())
 	}
 	if !data.GroupPolicyId.IsNull() {
 		body, _ = sjson.Set(body, "groupPolicyId", data.GroupPolicyId.ValueString())
@@ -150,22 +122,6 @@ func (data ResourceApplianceVLANsItems) toBody(ctx context.Context, state Resour
 	}
 	if !data.TemplateVlanType.IsNull() {
 		body, _ = sjson.Set(body, "templateVlanType", data.TemplateVlanType.ValueString())
-	}
-	if !data.VpnNatSubnet.IsNull() {
-		body, _ = sjson.Set(body, "vpnNatSubnet", data.VpnNatSubnet.ValueString())
-	}
-	if len(data.FixedIpAssignments) > 0 {
-		body, _ = sjson.Set(body, "fixedIpAssignments", map[string]interface{}{})
-		for key, item := range data.FixedIpAssignments {
-			itemBody := ""
-			if !item.Ip.IsNull() {
-				itemBody, _ = sjson.Set(itemBody, "ip", item.Ip.ValueString())
-			}
-			if !item.Name.IsNull() {
-				itemBody, _ = sjson.Set(itemBody, "name", item.Name.ValueString())
-			}
-			body, _ = sjson.SetRaw(body, "fixedIpAssignments."+key, itemBody)
-		}
 	}
 	if !data.Ipv6Enabled.IsNull() {
 		body, _ = sjson.Set(body, "ipv6.enabled", data.Ipv6Enabled.ValueBool())
@@ -213,27 +169,6 @@ func (data ResourceApplianceVLANsItems) toBody(ctx context.Context, state Resour
 			body, _ = sjson.SetRaw(body, "dhcpOptions.-1", itemBody)
 		}
 	}
-	if !data.DhcpRelayServerIps.IsNull() {
-		var values []string
-		data.DhcpRelayServerIps.ElementsAs(ctx, &values, false)
-		body, _ = sjson.Set(body, "dhcpRelayServerIps", values)
-	}
-	if len(data.ReservedIpRanges) > 0 {
-		body, _ = sjson.Set(body, "reservedIpRanges", []interface{}{})
-		for _, item := range data.ReservedIpRanges {
-			itemBody := ""
-			if !item.Comment.IsNull() {
-				itemBody, _ = sjson.Set(itemBody, "comment", item.Comment.ValueString())
-			}
-			if !item.End.IsNull() {
-				itemBody, _ = sjson.Set(itemBody, "end", item.End.ValueString())
-			}
-			if !item.Start.IsNull() {
-				itemBody, _ = sjson.Set(itemBody, "start", item.Start.ValueString())
-			}
-			body, _ = sjson.SetRaw(body, "reservedIpRanges.-1", itemBody)
-		}
-	}
 	return body
 }
 
@@ -259,16 +194,6 @@ func (data *ResourceApplianceVLANs) fromBody(ctx context.Context, res meraki.Res
 		} else {
 			data.Cidr = types.StringNull()
 		}
-		if value := res.Get("dhcpBootFilename"); value.Exists() && value.Value() != nil {
-			data.DhcpBootFilename = types.StringValue(value.String())
-		} else {
-			data.DhcpBootFilename = types.StringNull()
-		}
-		if value := res.Get("dhcpBootNextServer"); value.Exists() && value.Value() != nil {
-			data.DhcpBootNextServer = types.StringValue(value.String())
-		} else {
-			data.DhcpBootNextServer = types.StringNull()
-		}
 		if value := res.Get("dhcpBootOptionsEnabled"); value.Exists() && value.Value() != nil {
 			data.DhcpBootOptionsEnabled = types.BoolValue(value.Bool())
 		} else {
@@ -283,11 +208,6 @@ func (data *ResourceApplianceVLANs) fromBody(ctx context.Context, res meraki.Res
 			data.DhcpLeaseTime = types.StringValue(value.String())
 		} else {
 			data.DhcpLeaseTime = types.StringNull()
-		}
-		if value := res.Get("dnsNameservers"); value.Exists() && value.Value() != nil {
-			data.DnsNameservers = types.StringValue(value.String())
-		} else {
-			data.DnsNameservers = types.StringNull()
 		}
 		if value := res.Get("groupPolicyId"); value.Exists() && value.Value() != nil {
 			data.GroupPolicyId = types.StringValue(value.String())
@@ -318,30 +238,6 @@ func (data *ResourceApplianceVLANs) fromBody(ctx context.Context, res meraki.Res
 			data.TemplateVlanType = types.StringValue(value.String())
 		} else {
 			data.TemplateVlanType = types.StringNull()
-		}
-		if value := res.Get("vpnNatSubnet"); value.Exists() && value.Value() != nil {
-			data.VpnNatSubnet = types.StringValue(value.String())
-		} else {
-			data.VpnNatSubnet = types.StringNull()
-		}
-		if value := res.Get("fixedIpAssignments"); value.Exists() && value.Value() != nil {
-			data.FixedIpAssignments = make(map[string]ResourceApplianceVLANsFixedIpAssignments)
-			value.ForEach(func(k, res gjson.Result) bool {
-				parent := &data
-				data := ResourceApplianceVLANsFixedIpAssignments{}
-				if value := res.Get("ip"); value.Exists() && value.Value() != nil {
-					data.Ip = types.StringValue(value.String())
-				} else {
-					data.Ip = types.StringNull()
-				}
-				if value := res.Get("name"); value.Exists() && value.Value() != nil {
-					data.Name = types.StringValue(value.String())
-				} else {
-					data.Name = types.StringNull()
-				}
-				(*parent).FixedIpAssignments[k.String()] = data
-				return true
-			})
 		}
 		if value := res.Get("ipv6.enabled"); value.Exists() && value.Value() != nil {
 			data.Ipv6Enabled = types.BoolValue(value.Bool())
@@ -411,35 +307,6 @@ func (data *ResourceApplianceVLANs) fromBody(ctx context.Context, res meraki.Res
 				return true
 			})
 		}
-		if value := res.Get("dhcpRelayServerIps"); value.Exists() && value.Value() != nil {
-			data.DhcpRelayServerIps = helpers.GetStringList(value.Array())
-		} else {
-			data.DhcpRelayServerIps = types.ListNull(types.StringType)
-		}
-		if value := res.Get("reservedIpRanges"); value.Exists() && value.Value() != nil {
-			data.ReservedIpRanges = make([]ResourceApplianceVLANsReservedIpRanges, 0)
-			value.ForEach(func(k, res gjson.Result) bool {
-				parent := &data
-				data := ResourceApplianceVLANsReservedIpRanges{}
-				if value := res.Get("comment"); value.Exists() && value.Value() != nil {
-					data.Comment = types.StringValue(value.String())
-				} else {
-					data.Comment = types.StringNull()
-				}
-				if value := res.Get("end"); value.Exists() && value.Value() != nil {
-					data.End = types.StringValue(value.String())
-				} else {
-					data.End = types.StringNull()
-				}
-				if value := res.Get("start"); value.Exists() && value.Value() != nil {
-					data.Start = types.StringValue(value.String())
-				} else {
-					data.Start = types.StringNull()
-				}
-				(*parent).ReservedIpRanges = append((*parent).ReservedIpRanges, data)
-				return true
-			})
-		}
 		(*parent).Items = append((*parent).Items, data)
 		return true
 	})
@@ -496,16 +363,6 @@ func (data *ResourceApplianceVLANs) fromBodyPartial(ctx context.Context, res mer
 		} else {
 			data.Cidr = types.StringNull()
 		}
-		if value := res.Get("dhcpBootFilename"); value.Exists() && !data.DhcpBootFilename.IsNull() {
-			data.DhcpBootFilename = types.StringValue(value.String())
-		} else {
-			data.DhcpBootFilename = types.StringNull()
-		}
-		if value := res.Get("dhcpBootNextServer"); value.Exists() && !data.DhcpBootNextServer.IsNull() {
-			data.DhcpBootNextServer = types.StringValue(value.String())
-		} else {
-			data.DhcpBootNextServer = types.StringNull()
-		}
 		if value := res.Get("dhcpBootOptionsEnabled"); value.Exists() && !data.DhcpBootOptionsEnabled.IsNull() {
 			data.DhcpBootOptionsEnabled = types.BoolValue(value.Bool())
 		} else {
@@ -520,11 +377,6 @@ func (data *ResourceApplianceVLANs) fromBodyPartial(ctx context.Context, res mer
 			data.DhcpLeaseTime = types.StringValue(value.String())
 		} else {
 			data.DhcpLeaseTime = types.StringNull()
-		}
-		if value := res.Get("dnsNameservers"); value.Exists() && !data.DnsNameservers.IsNull() {
-			data.DnsNameservers = types.StringValue(value.String())
-		} else {
-			data.DnsNameservers = types.StringNull()
 		}
 		if value := res.Get("groupPolicyId"); value.Exists() && !data.GroupPolicyId.IsNull() {
 			data.GroupPolicyId = types.StringValue(value.String())
@@ -555,28 +407,6 @@ func (data *ResourceApplianceVLANs) fromBodyPartial(ctx context.Context, res mer
 			data.TemplateVlanType = types.StringValue(value.String())
 		} else {
 			data.TemplateVlanType = types.StringNull()
-		}
-		if value := res.Get("vpnNatSubnet"); value.Exists() && !data.VpnNatSubnet.IsNull() {
-			data.VpnNatSubnet = types.StringValue(value.String())
-		} else {
-			data.VpnNatSubnet = types.StringNull()
-		}
-		for i, item := range data.FixedIpAssignments {
-			parent := &data
-			data := item
-			parentRes := &res
-			res := parentRes.Get(fmt.Sprintf("fixedIpAssignments.%s", i))
-			if value := res.Get("ip"); value.Exists() && !data.Ip.IsNull() {
-				data.Ip = types.StringValue(value.String())
-			} else {
-				data.Ip = types.StringNull()
-			}
-			if value := res.Get("name"); value.Exists() && !data.Name.IsNull() {
-				data.Name = types.StringValue(value.String())
-			} else {
-				data.Name = types.StringNull()
-			}
-			(*parent).FixedIpAssignments[i] = data
 		}
 		if value := res.Get("ipv6.enabled"); value.Exists() && !data.Ipv6Enabled.IsNull() {
 			data.Ipv6Enabled = types.BoolValue(value.Bool())
@@ -680,64 +510,6 @@ func (data *ResourceApplianceVLANs) fromBodyPartial(ctx context.Context, res mer
 			}
 			(*parent).DhcpOptions[i] = data
 		}
-		if value := res.Get("dhcpRelayServerIps"); value.Exists() && !data.DhcpRelayServerIps.IsNull() {
-			data.DhcpRelayServerIps = helpers.GetStringList(value.Array())
-		} else {
-			data.DhcpRelayServerIps = types.ListNull(types.StringType)
-		}
-		for i := 0; i < len(data.ReservedIpRanges); i++ {
-			keys := [...]string{"comment", "end", "start"}
-			keyValues := [...]string{data.ReservedIpRanges[i].Comment.ValueString(), data.ReservedIpRanges[i].End.ValueString(), data.ReservedIpRanges[i].Start.ValueString()}
-
-			parent := &data
-			data := (*parent).ReservedIpRanges[i]
-			parentRes := &res
-			var res gjson.Result
-
-			parentRes.Get("reservedIpRanges").ForEach(
-				func(_, v gjson.Result) bool {
-					found := false
-					for ik := range keys {
-						if v.Get(keys[ik]).String() != keyValues[ik] {
-							found = false
-							break
-						}
-						found = true
-					}
-					if found {
-						res = v
-						return false
-					}
-					return true
-				},
-			)
-			if !res.Exists() {
-				tflog.Debug(ctx, fmt.Sprintf("removing ReservedIpRanges[%d] = %+v",
-					i,
-					(*parent).ReservedIpRanges[i],
-				))
-				(*parent).ReservedIpRanges = slices.Delete((*parent).ReservedIpRanges, i, i+1)
-				i--
-
-				continue
-			}
-			if value := res.Get("comment"); value.Exists() && !data.Comment.IsNull() {
-				data.Comment = types.StringValue(value.String())
-			} else {
-				data.Comment = types.StringNull()
-			}
-			if value := res.Get("end"); value.Exists() && !data.End.IsNull() {
-				data.End = types.StringValue(value.String())
-			} else {
-				data.End = types.StringNull()
-			}
-			if value := res.Get("start"); value.Exists() && !data.Start.IsNull() {
-				data.Start = types.StringValue(value.String())
-			} else {
-				data.Start = types.StringNull()
-			}
-			(*parent).ReservedIpRanges[i] = data
-		}
 		(*parent).Items[i] = data
 	}
 	for i := len(toBeDeleted) - 1; i >= 0; i-- {
@@ -795,16 +567,6 @@ func (data *ResourceApplianceVLANs) fromBodyImport(ctx context.Context, res mera
 		} else {
 			data.Cidr = types.StringNull()
 		}
-		if value := res.Get("dhcpBootFilename"); value.Exists() && value.Value() != nil {
-			data.DhcpBootFilename = types.StringValue(value.String())
-		} else {
-			data.DhcpBootFilename = types.StringNull()
-		}
-		if value := res.Get("dhcpBootNextServer"); value.Exists() && value.Value() != nil {
-			data.DhcpBootNextServer = types.StringValue(value.String())
-		} else {
-			data.DhcpBootNextServer = types.StringNull()
-		}
 		if value := res.Get("dhcpBootOptionsEnabled"); value.Exists() && value.Value() != nil {
 			data.DhcpBootOptionsEnabled = types.BoolValue(value.Bool())
 		} else {
@@ -819,11 +581,6 @@ func (data *ResourceApplianceVLANs) fromBodyImport(ctx context.Context, res mera
 			data.DhcpLeaseTime = types.StringValue(value.String())
 		} else {
 			data.DhcpLeaseTime = types.StringNull()
-		}
-		if value := res.Get("dnsNameservers"); value.Exists() && value.Value() != nil {
-			data.DnsNameservers = types.StringValue(value.String())
-		} else {
-			data.DnsNameservers = types.StringNull()
 		}
 		if value := res.Get("groupPolicyId"); value.Exists() && value.Value() != nil {
 			data.GroupPolicyId = types.StringValue(value.String())
@@ -854,30 +611,6 @@ func (data *ResourceApplianceVLANs) fromBodyImport(ctx context.Context, res mera
 			data.TemplateVlanType = types.StringValue(value.String())
 		} else {
 			data.TemplateVlanType = types.StringNull()
-		}
-		if value := res.Get("vpnNatSubnet"); value.Exists() && value.Value() != nil {
-			data.VpnNatSubnet = types.StringValue(value.String())
-		} else {
-			data.VpnNatSubnet = types.StringNull()
-		}
-		if value := res.Get("fixedIpAssignments"); value.Exists() && value.Value() != nil && len(value.Array()) > 0 {
-			data.FixedIpAssignments = make(map[string]ResourceApplianceVLANsFixedIpAssignments)
-			value.ForEach(func(k, res gjson.Result) bool {
-				parent := &data
-				data := ResourceApplianceVLANsFixedIpAssignments{}
-				if value := res.Get("ip"); value.Exists() && value.Value() != nil {
-					data.Ip = types.StringValue(value.String())
-				} else {
-					data.Ip = types.StringNull()
-				}
-				if value := res.Get("name"); value.Exists() && value.Value() != nil {
-					data.Name = types.StringValue(value.String())
-				} else {
-					data.Name = types.StringNull()
-				}
-				(*parent).FixedIpAssignments[k.String()] = data
-				return true
-			})
 		}
 		if value := res.Get("ipv6.enabled"); value.Exists() && value.Value() != nil {
 			data.Ipv6Enabled = types.BoolValue(value.Bool())
@@ -947,35 +680,6 @@ func (data *ResourceApplianceVLANs) fromBodyImport(ctx context.Context, res mera
 				return true
 			})
 		}
-		if value := res.Get("dhcpRelayServerIps"); value.Exists() && value.Value() != nil && len(value.Array()) > 0 {
-			data.DhcpRelayServerIps = helpers.GetStringList(value.Array())
-		} else {
-			data.DhcpRelayServerIps = types.ListNull(types.StringType)
-		}
-		if value := res.Get("reservedIpRanges"); value.Exists() && value.Value() != nil && len(value.Array()) > 0 {
-			data.ReservedIpRanges = make([]ResourceApplianceVLANsReservedIpRanges, 0)
-			value.ForEach(func(k, res gjson.Result) bool {
-				parent := &data
-				data := ResourceApplianceVLANsReservedIpRanges{}
-				if value := res.Get("comment"); value.Exists() && value.Value() != nil {
-					data.Comment = types.StringValue(value.String())
-				} else {
-					data.Comment = types.StringNull()
-				}
-				if value := res.Get("end"); value.Exists() && value.Value() != nil {
-					data.End = types.StringValue(value.String())
-				} else {
-					data.End = types.StringNull()
-				}
-				if value := res.Get("start"); value.Exists() && value.Value() != nil {
-					data.Start = types.StringValue(value.String())
-				} else {
-					data.Start = types.StringNull()
-				}
-				(*parent).ReservedIpRanges = append((*parent).ReservedIpRanges, data)
-				return true
-			})
-		}
 		(*parent).Items[i] = data
 	}
 	for i := len(toBeDeleted) - 1; i >= 0; i-- {
@@ -1020,12 +724,6 @@ func (data *ResourceApplianceVLANs) hasChanges(ctx context.Context, state *Resou
 	if !item.Cidr.Equal(stateItem.Cidr) {
 		hasChanges = true
 	}
-	if !item.DhcpBootFilename.Equal(stateItem.DhcpBootFilename) {
-		hasChanges = true
-	}
-	if !item.DhcpBootNextServer.Equal(stateItem.DhcpBootNextServer) {
-		hasChanges = true
-	}
 	if !item.DhcpBootOptionsEnabled.Equal(stateItem.DhcpBootOptionsEnabled) {
 		hasChanges = true
 	}
@@ -1033,9 +731,6 @@ func (data *ResourceApplianceVLANs) hasChanges(ctx context.Context, state *Resou
 		hasChanges = true
 	}
 	if !item.DhcpLeaseTime.Equal(stateItem.DhcpLeaseTime) {
-		hasChanges = true
-	}
-	if !item.DnsNameservers.Equal(stateItem.DnsNameservers) {
 		hasChanges = true
 	}
 	if !item.GroupPolicyId.Equal(stateItem.GroupPolicyId) {
@@ -1054,12 +749,6 @@ func (data *ResourceApplianceVLANs) hasChanges(ctx context.Context, state *Resou
 		hasChanges = true
 	}
 	if !item.TemplateVlanType.Equal(stateItem.TemplateVlanType) {
-		hasChanges = true
-	}
-	if !item.VpnNatSubnet.Equal(stateItem.VpnNatSubnet) {
-		hasChanges = true
-	}
-	if !maps.Equal(item.FixedIpAssignments, stateItem.FixedIpAssignments) {
 		hasChanges = true
 	}
 	if !item.Ipv6Enabled.Equal(stateItem.Ipv6Enabled) {
@@ -1100,24 +789,6 @@ func (data *ResourceApplianceVLANs) hasChanges(ctx context.Context, state *Resou
 				hasChanges = true
 			}
 			if !item.DhcpOptions[i].Value.Equal(stateItem.DhcpOptions[i].Value) {
-				hasChanges = true
-			}
-		}
-	}
-	if !item.DhcpRelayServerIps.Equal(stateItem.DhcpRelayServerIps) {
-		hasChanges = true
-	}
-	if len(item.ReservedIpRanges) != len(stateItem.ReservedIpRanges) {
-		hasChanges = true
-	} else {
-		for i := range item.ReservedIpRanges {
-			if !item.ReservedIpRanges[i].Comment.Equal(stateItem.ReservedIpRanges[i].Comment) {
-				hasChanges = true
-			}
-			if !item.ReservedIpRanges[i].End.Equal(stateItem.ReservedIpRanges[i].End) {
-				hasChanges = true
-			}
-			if !item.ReservedIpRanges[i].Start.Equal(stateItem.ReservedIpRanges[i].Start) {
 				hasChanges = true
 			}
 		}