+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 - Add support for `both` value to `ip_version` attribute of `wireless_ssid_l3_firewall_rules` resource
 - Fix issue with configuring DHCP relay (`dhcp_handling`, `dhcp_relay_server_ips` atributes) in `meraki_appliance_vlan_dhcp` resource, [link](https://github.com/CiscoDevNet/terraform-provider-meraki/pull/123)
 - Add `fixed_ip_assignments` and `vpn_nat_subnet` attributes to `meraki_appliance_vlan_dhcp` resource, [link](https://github.com/CiscoDevNet/terraform-provider-meraki/pull/124)
+- BREAKING CHANGE: Remove PUT-only attributes of `meraki_appliance_vlan` (present in `meraki_appliance_vlan_dhcp` resource), [link](https://github.com/CiscoDevNet/terraform-provider-meraki/pull/124):
+  - `dhcp_boot_filename`,
+  - `dhcp_boot_next_server`,
+  - `dns_nameservers`,
+  - `vpn_nat_subnet`,
+  - `fixed_ip_assignments`,
+  - `dhcp_relay_server_ips`.
 
 ## 1.7.1